+++ CHANGELOG.md
@@ -1,5 +1,16 @@
-## [0.1.1] - 2025-03-06
+## [0.2.2] - 2025-03-07
 ### Added
+- Github Actions Unittest
+- Unittest of Filter and Logger modules
+
+## [0.2.1] - 2025-03-07
+### Added
+- Github Actions Pylint
+- Github Actions Docker
+- Creation of ProxyServer class
+
+## [0.2.0] - 2025-03-06
+### Patch
 - Editing page 403.
 - Readme update.